(refactor) Rank enum extract
