+++ src/test/java/lotto/domain/WinningNumbersTest.java
@@ -7,7 +7,6 @@ import static org.junit.jupiter.api.Assertions.assertAll;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
-import lotto.domain.WinningNumbers.Rank;
 import org.junit.jupiter.api.Test;
 
 public class WinningNumbersTest {
