+++ src/test/java/lotto/domain/WinningPercentTest.java
@@ -4,7 +4,6 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.EnumMap;
 import java.util.Map;
-import lotto.domain.WinningNumbers.Rank;
 import org.junit.jupiter.api.Test;
 
 public class WinningPercentTest {
